+++ src/types/addToCartCountType.ts
@@ -1,0 +1,4 @@
+export interface AddToCartCountProps {
+  id: number;
+  onDeleteCart: () => void;
+}
